backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
alembic
